+++ tests/test_parsing.py
@@ -215,6 +215,7 @@ def test_track_artists(artists, expected):
         ("Track *digital-only", True, "Track"),
         ("DIGITAL 2. Track", True, "Track"),
         ("Track (digital)", True, "Track"),
+        ("Bonus : Track", True, "Track"),
     ],
 )
 def test_check_digital_only(name, expected_digital_only, expected_name):